a more concise way to return early when the post doesn't need to be converted
using return() in switch() will trigger a bogus NOTE:

* checking R code for possible problems ... [8s] NOTE
Possibly missing '()' after 'return' in the following function:
  'hugo_convert_one'

this is because codetools::findGlobals() will identify `return` as a global variable

+++ R/hugo.R
@@ -332,9 +332,7 @@ hugo_convert_one = function(file, to = c('YAML', 'TOML', 'JSON')) {
   }
   x = x[x != ''][1]
   to = match.arg(to)
-  switch(
-    to, YAML = if (x == '---') return(), TOML = if (x == '+++') return, JSON = if (x == '{') return()
-  )
+  if (x == c(YAML = '---', TOML = '+++', JSON = '{')[to]) return()
   file = normalizePath(file)
   tmp = tempfile(); on.exit(unlink(tmp, recursive = TRUE), add = TRUE)
   dir.create(tmp)
